refactorisation des controlleurs
