+++ app/config/config.yml
@@ -110,7 +110,7 @@ ensepar_html2pdf:
     margin: [50,50,50,50]
 
 fos_user:
-    db_driver: orm # other valid values are 'mongodb' and 'couchdb'
+    db_driver: orm
     firewall_name: main
     user_class: UserBundle\Entity\User
     from_email:
