+++ composer.json
@@ -21,7 +21,7 @@
   },
   "require": {
     "php": ">=5.5.9",
-    "symfony/symfony": "3.2.*",
+    "symfony/symfony": "~3.2",
     "doctrine/orm": "^2.5",
     "doctrine/doctrine-bundle": "^1.6",
     "doctrine/doctrine-cache-bundle": "^1.2",
